backend/pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ecpm"
version = "0.1.0"
description = "Economic Crisis Prediction Model - Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.30,<1.0",
    "sqlalchemy[asyncio]>=2.0,<3.0",
    "asyncpg>=0.29,<1.0",
    "alembic>=1.13,<2.0",
    "celery[redis]>=5.4,<6.0",
    "redis[hiredis]>=5.0,<6.0",
    "fredapi>=0.5,<1.0",
    "beaapi>=0.0.4,<1.0",
    "tenacity>=8.2,<10.0",
    "pydantic-settings>=2.0,<3.0",
    "structlog>=24.0,<26.0",
    "pandas>=2.2,<3.0",
    "pyyaml>=6.0,<7.0",
    "httpx>=0.27,<1.0",
    "sse-starlette>=2.0,<3.0",
    "statsmodels>=0.14.6,<1.0",
    "scikit-learn>=1.4,<2.0",
    "python-jose[cryptography]>=3.3,<4.0",
    "bcrypt>=4.0,<6.0",
    "python-multipart>=0.0.9,<1.0",
    "slowapi>=0.1.9,<1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.24,<1.0",
    "pytest-timeout>=2.3,<3.0",
    "httpx>=0.27,<1.0",
    "pip-audit>=2.7",
    "safety>=3.0",
]

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
timeout = 30
testpaths = ["tests"]
